Refactor #398 - clean up pref package
