+++ core/data-di/src/main/kotlin/com/nlab/reminder/core/data/di/AppScopeDataModule.kt
@@ -43,7 +43,7 @@ import com.nlab.reminder.core.local.database.dao.ScheduleDAO
 import com.nlab.reminder.core.local.database.dao.ScheduleTagListDAO
 import com.nlab.reminder.core.local.database.dao.TagRelationDAO
 import com.nlab.reminder.core.local.database.dao.TagDAO
-import com.nlab.reminder.core.local.datastore.PreferenceDataSource
+import com.nlab.reminder.core.local.datastore.preference.PreferenceDataSource
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
