+++ core/local/build.gradle.kts
@@ -15,7 +15,7 @@ dependencies {
     implementation(projects.core.kotlin)
 
     api(libs.androidx.datastore.preferences)
-    api(libs.androidx.room.runtime)
     api(libs.androidx.room.ktx)
     ksp(libs.androidx.room.compiler)
+    api(libs.kotlinx.datetime)
 }
