+++ core/local/src/main/kotlin/com/nlab/reminder/core/local/datastore/configuration/ReminderPreferenceDataStore.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 The N's lab Open Source Project
+ * Copyright (C) 2025 The N's lab Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 @file:Suppress("FunctionName")
 
-package com.nlab.reminder.core.local.datastore
+package com.nlab.reminder.core.local.datastore.configuration
 
 import android.content.Context
 import androidx.datastore.core.DataStore
@@ -28,7 +28,7 @@ import kotlinx.coroutines.CoroutineScope
 /**
  * @author Doohyun
  */
-fun ReminderDataStore(
+fun ReminderPreferenceDataStore(
     context: Context,
     coroutineScope: CoroutineScope
 ): DataStore<Preferences> = PreferenceDataStoreFactory.create(
