+++ core/local/src/main/kotlin/com/nlab/reminder/core/local/datastore/preference/PreferenceDataSource.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 The N's lab Open Source Project
+ * Copyright (C) 2025 The N's lab Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,30 +14,24 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.local.datastore
+package com.nlab.reminder.core.local.datastore.preference
 
 import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.MutablePreferences
 import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.edit
 import com.nlab.reminder.core.kotlin.Result
-import com.nlab.reminder.core.kotlin.catching
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 
 /**
  * @author Doohyun
  */
-internal suspend fun DataStore<Preferences>.tryEdit(
-    transform: suspend (MutablePreferences) -> Unit
-): Result<Unit> = catching { edit(transform) }
+class PreferenceDataSource(private val dataStore: DataStore<Preferences>) {
+    fun getAllScheduleCompleteShownAsStream(): Flow<Boolean> = dataStore.getAsStream(
+        key = PreferenceKeys.ALL_SCHEDULE_COMPLETE_SHOWN,
+        defaultValue = false
+    )
 
-internal suspend fun <T> DataStore<Preferences>.tryEdit(
-    key: Preferences.Key<T>,
-    value: T
-): Result<Unit> = tryEdit { preferences -> preferences[key] = value }
-
-internal fun <T> DataStore<Preferences>.getAsStream(
-    key: Preferences.Key<T>,
-    defaultValue: T
-): Flow<T> = data.map { it[key] ?: defaultValue }
+    suspend fun setAllScheduleCompleteShown(value: Boolean): Result<Unit> = dataStore.tryEdit(
+        key = PreferenceKeys.ALL_SCHEDULE_COMPLETE_SHOWN,
+        value
+    )
+}
